requirements.txt
amqp==5.1.1
asgiref==3.7.2
autopep8==2.0.1
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.4
django-environ==0.10.0
djangorestframework==3.14.0
drf-yasg==1.21.7
future==0.18.2
inflection==0.5.1
kombu==5.3.2
mysqlclient==2.2.0
numpy==1.22.2
opencv-python==4.7.0.72
packaging==23.1
pandas==1.5.2
prompt-toolkit==3.0.39
pycodestyle==2.10.0
PyPDF3==1.0.6
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tensorboard==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
tf-estimator-nightly==2.8.0.dev2021122109
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.6
yapf==0.32.0
